Feat: idea and todo input
