+++ src/components/Input/IdeaInput.tsx
@@ -9,6 +9,7 @@ import { observer } from "mobx-react-lite";
 import { useEffect, useRef, useState } from "react";
 import { IoIosAdd } from "react-icons/io";
 import { cardStore } from "../../store/cardStore";
+import { getFilteredTags } from "../../utils/input";
 
 const tabs = [
   { label: "靈感", id: "idea" },
@@ -18,11 +19,16 @@ const tabs = [
 export const IdeaInput = observer(() => {
   const inputRef = useRef<HTMLInputElement>(null);
 
+  const [selectedTab, setSelectedTab] = useState<"idea" | "todo">("idea");
   const [input, setInput] = useState<string>("");
   const [tagInput, setTagInput] = useState<string>("");
-  const [selectedTab, setSelectedTab] = useState<"idea" | "todo">("idea");
+
+  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setInput(e.target.value);
+  };
 
   const handleAddCard = () => {
+    console.log("handleAddCard");
     if (!input) return;
     cardStore.addCard(selectedTab, input, [tagInput]);
     setInput("");
@@ -33,11 +39,13 @@ export const IdeaInput = observer(() => {
     key === "idea" ? setSelectedTab("todo") : setSelectedTab("idea");
   };
 
-  const handleKeyDown = (e: KeyboardEvent) => {
+  const handleKeyDown = (
+    e: KeyboardEvent | React.KeyboardEvent<HTMLInputElement>,
+  ) => {
     if (e.key === "Escape") inputRef.current?.blur();
     if (e.key === "Enter") handleAddCard();
-    if (e.key === "i" && e.metaKey && e.shiftKey) handleTab(selectedTab);
     if (e.key === "i" && e.metaKey) inputRef.current?.focus();
+    if (e.key === "i" && e.metaKey && e.shiftKey) handleTab(selectedTab);
   };
 
   useEffect(() => {
@@ -47,21 +55,12 @@ export const IdeaInput = observer(() => {
     };
   }, [selectedTab]);
 
-  const getFilteredTags = (tagInput: string) => {
-    const tags = cardStore.getAllTags();
-    const emptyTagPlaceholder = [`新增 ${tagInput}`];
-    const filteredTags = tags.filter(
-      (tag) =>
-        tag.toLowerCase().includes(tagInput.toLowerCase()) && tag.length > 0,
-    );
-
-    if (filteredTags.length === 0) return emptyTagPlaceholder;
-    return filteredTags;
-  };
-
   const filteredTags = getFilteredTags(tagInput);
   const placeholder =
     selectedTab === "idea" ? "捕捉您的靈感..." : "新增待辦...";
+  const AddIcon = (
+    <IoIosAdd className="cursor-pointer" onClick={handleAddCard} />
+  );
 
   return (
     <section className="mb-8 flex flex-col items-center">
@@ -74,33 +73,31 @@ export const IdeaInput = observer(() => {
       >
         {(item) => <Tab key={item.id} title={item.label} />}
       </Tabs>
-      <div className="flex items-center gap-4">
+      <form className="flex items-center gap-4">
         <Input
-          size="sm"
           value={input}
           name={selectedTab}
           ref={inputRef}
-          endContent={
-            <IoIosAdd className="cursor-pointer" onClick={handleAddCard} />
-          }
+          size="sm"
           placeholder={placeholder}
-          onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
-            setInput(e.target.value)
-          }
+          endContent={AddIcon}
+          onChange={handleInputChange}
         />
         <Autocomplete
-          className="w-30"
-          variant="bordered"
           label="標籤"
-          size="sm"
+          variant="bordered"
+          inputValue={tagInput}
           allowsCustomValue
-          onInputChange={setTagInput}
+          className="w-25"
+          size="sm"
+          onInputChange={(input) => setTagInput(input)}
+          onKeyDown={handleKeyDown}
         >
-          {filteredTags.map((tag, index) => (
-            <AutocompleteItem key={index}>{tag}</AutocompleteItem>
+          {filteredTags.map((tag) => (
+            <AutocompleteItem key={tag}>{tag}</AutocompleteItem>
           ))}
         </Autocomplete>
-      </div>
+      </form>
     </section>
   );
 });
