+++ src/components/Input/index.tsx
@@ -1,0 +1,1 @@
+export { IdeaInput } from "./IdeaInput";
