+++ src/components/Section/index.tsx
@@ -1,4 +1,3 @@
-export { IdeaInput } from "./IdeaInput";
 export { TodayIdea } from "./TodayIdea";
 export { TodayReminder } from "./TodayReminder";
 export { TodayTodo } from "./TodayTodo";
